refactor: Consolidate test files in one place.
